refactor(parquet): review round 2 comment + naming cleanups
- Rename NewStreamBuffer's maxSize -> sizeHint (it's the initial size, not a
  hard cap; the buffer still grows for an oversized value).
- Rename setStreamingValues/drainAndClose limit -> rawSrc.
- Move streamablePhysicalType/streamableCodec to row_group_reader.go (their
  only caller).
- Reword the Fill lifetime contract, the V2 level-read comment, a decoder-test
  comment, and trim the EnablePageStreaming doc.
- Rename a test's shadowing `const cap` to batchSize.
